Follow-up refactoring of loggers
Updates in Logger:
    - Use map instead of vector for Logger::_properties, to allow for named access.
    - Use vector instead of initializer_list in constructors.
    - Add an empty constructor (for logger::Combine).

- Port logger::Combine.
- Port logger::EAH (and rename tests).
- Better naming scheme for headers.
- Move interface within the header with the same name.
- Adds correct legends in test_eah-distrib.
