+++ include/ioh/logger.hpp
@@ -1,8 +1,9 @@
 #pragma once
 
 #include "logger/loginfo.hpp"
-#include "logger/api.hpp"
 #include "logger/properties.hpp"
-#include "logger/watchers.hpp"
 #include "logger/triggers.hpp"
-
+#include "logger/loggers.hpp"
+#include "logger/store.hpp"
+#include "logger/combine.hpp"
+#include "logger/eah.hpp"
