+++ tests/cpp/logger/test_eah.cpp
@@ -12,7 +12,7 @@ TEST(eaf, bbob)
     auto sample_size = 100;
 
     ioh::suite::BBOB suite({1, 2}, {1, 2}, {2, 10});
-    ioh::logger::EAF logger(0, 6e7, 20, 0, sample_size, 20);
+    ioh::logger::EAH logger(0, 6e7, 20, 0, sample_size, 20);
 
     suite.attach_logger(logger);
 
@@ -21,15 +21,12 @@ TEST(eaf, bbob)
         2554, 2785, 2850, 2915, 2982, 3046, 3148, 3242
     };
 
-
-    for (const auto &p : suite)
-    {
-        for (auto r = 0; r < 2; r++)
-        {
+    for (const auto &p : suite) {
+        for (auto r = 0; r < 2; r++) {
             for (auto s = 0; s < sample_size; ++s)
                 (*p)(ioh::common::Random::uniform(p->meta_data().n_variables));
 
-            EXPECT_EQ(ioh::logger::eaf::stat::sum(logger), attainments_sum.front());
+            EXPECT_EQ(ioh::logger::eah::stat::sum(logger), attainments_sum.front());
             p->reset();
             attainments_sum.pop_front();
         }
